Move map.component, make lat/lng double, fix routes, create factory
