+++ metka-ui/src/app/app-routing.module.ts
@@ -4,11 +4,6 @@ import { RouterModule, Routes } from '@angular/router';
 const routes: Routes = [
   {
     path: '',
-    redirectTo: 'home',
-    pathMatch: 'full',
-  },
-  {
-    path: 'home',
     loadChildren: () =>
       import('./features/home-page/home-page.module').then(m => m.HomePageModule),
   },
@@ -16,6 +11,15 @@ const routes: Routes = [
     path: 'table',
     loadChildren: () =>
       import('./features/table-page/table-page.module').then(m => m.TablePageModule),
+  },
+  {
+    path: '',
+    redirectTo: 'home',
+    pathMatch: 'full',
+  },
+  {
+    path: '**',
+    redirectTo: '',
   },
 ];
