+++ metka-ui/src/app/features/table-page/table-page-routing.module.ts
@@ -2,10 +2,16 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { TablePageComponent } from './table-page.component';
 
-const routes: Routes = [{ path: '', component: TablePageComponent }];
+const routes: Routes = [
+  {
+    path: '',
+    component: TablePageComponent,
+  },
+];
 
 @NgModule({
   imports: [RouterModule.forChild(routes)],
-  exports: [RouterModule]
+  exports: [RouterModule],
 })
-export class TablePageRoutingModule { }
+export class TablePageRoutingModule {
+}
