+++ metka-ui/src/app/shared/components/map/map.component.html
