+++ metka-ui/src/app/shared/util/factory-helper.ts
@@ -3,12 +3,12 @@ import { DtoMarker } from '../../core/model/dto-marker';
 
 export class FactoryHelper {
 
-  public static newDtoMarker(marker: Marker, date: Date): DtoMarker {
+  public static newDtoMarker(marker: Marker, dateIso: string | null): DtoMarker {
     const id = marker.id ?? 0;
     const lat = marker.lat;
     const lng = marker.lng;
 
-    // return {id, lat, lng, date} as DtoMarker;
-    return {id: 0, lat, lng} as DtoMarker;
+    return {id: 0, lat, lng, mrkdate: dateIso} as DtoMarker;
+    // return {id: 0, lat, lng} as DtoMarker;
   }
 }
